// src/clients/node/src/types.ts
export enum TransferFlags {
  linked = (1 << 0),
  pending = (1 << 1),
  post_pending_transfer = (1 << 2),
  void_pending_transfer = (1 << 3),
  balancing_debit = (1 << 4),
  balancing_credit = (1 << 5),
  padding = (1 << 6),
}

export type Account = {
  id: bigint
  user_data: bigint
  reserved: Buffer
  ledger: number
  code: number
  flags: number
  debits_pending: bigint
  debits_posted: bigint
  credits_pending: bigint
  credits_posted: bigint
  timestamp: bigint
}

export type Transfer = {
  id: bigint
  debit_account_id: bigint
  credit_account_id: bigint
  user_data: bigint
  reserved: bigint
  pending_id: bigint
  timeout: bigint
  ledger: number
  code: number
  flags: number
  amount: bigint
  timestamp: bigint
}
// ...
export enum CreateTransferError {
  ok = 0,
  linked_event_failed = 1,
  linked_event_chain_open = 2,
  timestamp_must_be_zero = 3,
  reserved_flag = 4,
  reserved_field = 5,
  id_must_not_be_zero = 6,
  id_must_not_be_int_max = 7,
  flags_are_mutually_exclusive = 8,
  debit_account_id_must_not_be_zero = 9,
  debit_account_id_must_not_be_int_max = 10,
  credit_account_id_must_not_be_zero = 11,
  credit_account_id_must_not_be_int_max = 12,
  accounts_must_be_different = 13,
  pending_id_must_be_zero = 14,
  pending_id_must_not_be_zero = 15,
  pending_id_must_not_be_int_max = 16,
  pending_id_must_be_different = 17,
  timeout_reserved_for_pending_transfer = 18,
  ledger_must_not_be_zero = 19,
  code_must_not_be_zero = 20,
  amount_must_not_be_zero = 21,
  debit_account_not_found = 22,
  credit_account_not_found = 23,
  accounts_must_have_the_same_ledger = 24,
  transfer_must_have_the_same_ledger_as_accounts = 25,
  pending_transfer_not_found = 26,
  pending_transfer_not_pending = 27,
  pending_transfer_has_different_debit_account_id = 28,
  pending_transfer_has_different_credit_account_id = 29,
  pending_transfer_has_different_ledger = 30,
  pending_transfer_has_different_code = 31,
  exceeds_pending_transfer_amount = 32,
  pending_transfer_has_different_amount = 33,
  pending_transfer_already_posted = 34,
  pending_transfer_already_voided = 35,
  pending_transfer_expired = 36,
  exists_with_different_flags = 37,
  exists_with_different_debit_account_id = 38,
  exists_with_different_credit_account_id = 39,
  exists_with_different_pending_id = 40,
  exists_with_different_user_data = 41,
  exists_with_different_timeout = 42,
  exists_with_different_code = 43,
  exists_with_different_amount = 44,
  exists = 45,
  overflows_debits_pending = 46,
  overflows_credits_pending = 47,
  overflows_debits_posted = 48,
  overflows_credits_posted = 49,
  overflows_debits = 50,
  overflows_credits = 51,
  overflows_timeout = 52,
  exceeds_credits = 53,
  exceeds_debits = 54,
}

export type CreateAccountsError = {
  index: number
  result: CreateAccountError
}

export type CreateTransfersError = {
  index: number
  result: CreateTransferError
}

export enum Operation {
  reserved = 0,
  root = 1,
  register = 2,
  create_accounts = 3,
  create_transfers = 4,
  lookup_accounts = 5,
  lookup_transfers = 6,
}